Display maze with paint
Using the paint gem, a maze can be printed to a terminal screen as a
collection of colored blocks (an improvement over letters). Under the
new setup, walls are displayed as light blue (previously 'W'); regular
nodes are displayed as blank spaces (as before); the current path is
displayed in green (previously 'o'); and dead ends are displayed in red.
A dead end is a subset of the path that does not lead to the end of the
maze. Dead ends are displayed as green when originally traversed, and
turn red when the algorithm determines them to be dead ends. Change the
output method to erase the maze and redraw it at each step (as opposed
to simply clearing the screen by inserting empty lines).