Add ability to configure maze
Add maze_runner.rb, which takes options (in order): height, width,
display_time, and loops. Height and width specify the size of the maze.
Display time represents the time between each step of making and solving
the maze. Loops determines the number of mazes that will be displayed
before the script exits (defaulting to infinity). The script displays
the specified number of mazes, with the given configurations, one after
the other, erasing each maze before displaying the next one.