Add shortcuts to Maze for common nodes
Maze#start => The start of the maze. Currently the top, left node that
is not a wall.
Maze#entrance => The open wall between the start of the maze and the
outside of the maze. Currently located above the start.
Maze#finish => The end of the maze. Currently the bottom, right node
that is not a wall.
Maze#exit => The open wall between the end of the maze and the outside
of the maze. Currently located below the finish.