Refactor Builder class
Add Builder::path, which takes a maze and an optional block. It returns
the path generated when building the maze, and yields each node to the
block if given. Remove all logic for printing the maze and nodes from
Builder. Change MazeRunner to use Builder::path, using the block to
render the steps in creating the path.